Move QtChunkReceiver into experimental
